backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.14"
dependencies = [
    "django>=5.1",
    "django-ninja>=1.3",
    "dj-database-url>=2.3",
    "gunicorn>=23.0",
    "psycopg[binary]>=3.2",
    "whitenoise>=6.8",
]

[dependency-groups]
dev = [
    "beautifulsoup4>=4.12",
    "django-stubs>=5.2.9",
    "djlint>=1.36.4",
    "duckdb>=1.4.4",
    "pytest>=8.3",
    "pytest-django>=4.9",
    "requests>=2.32",
    "ruff>=0.8",
]
